+++ beaker/services/account.py
@@ -1,3 +1,5 @@
+from typing import List
+
 from cachetools import TTLCache, cached
 
 from ..data_model import *
@@ -5,9 +7,22 @@ from .service_client import ServiceClient
 
 
 class AccountClient(ServiceClient):
+    @property
+    def name(self):
+        """
+        A convenience property to get username of your Beaker account.
+        """
+        return self.whoami().name
+
     @cached(cache=TTLCache(maxsize=10, ttl=5 * 60))
     def whoami(self) -> Account:
         """
         Check who you are authenticated as.
         """
         return Account.from_json(self.request("user").json())
+
+    def list_organizations(self) -> List[Organization]:
+        """
+        List all organizations you are a member of.
+        """
+        return [Organization.from_json(d) for d in self.request("user/orgs").json()["data"]]